pyproject.toml
[project]
name = "instant-agents"
version = "0.1.0"
description = "Examples of how to build remarkably simple agents"
authors = [
    {name = "light magician",email = "email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "pydantic-ai[openai] (>=0.4.2,<0.5.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tavily-python (>=0.7.9,<0.8.0)"
]

[tool.poetry]
packages = [{include = "agent"}]
include = ["agent/system_prompt.txt"]

[project.scripts]
instant-agent = "agent.cli:main"


[build-system]
requires = ["setuptools>=45", "wheel", "poetry-core>=2.0.0,<3.0.0"]
build-backend = "setuptools.build_meta"
